bookbargains/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
Django==3.2.2
django-filter==2.4.0
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
